pyproject.toml
[project]
name = "engineering-hub"
version = "0.1.0"
description = "Agent-first workspace for acoustic engineering consulting"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jake Pfitsch"}
]
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "watchdog>=3.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "responses>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
mock-server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
engineering-hub = "engineering_hub.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=engineering_hub --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
